# app/api/enterprise/bulk-payments/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { z } from 'zod';

// Validation schemas
const bulkPaymentQuerySchema = z.object({
  status: z.enum(['draft', 'scheduled', 'processing', 'completed', 'failed', 'paused']).optional(),
  page: z.string().optional(),
  limit: z.string().optional(),
  sortBy: z.enum(['created', 'scheduled', 'amount', 'recipients']).optional(),
  sortOrder: z.enum(['asc', 'desc']).optional(),
  search: z.string().optional()
});

const createBulkPaymentSchema = z.object({
  name: z.string().min(1, "Batch name is required"),
  description: z.string().optional(),
  currency: z.string().length(3, "Currency must be 3 characters"),
  scheduledDate: z.string().datetime().optional(),
  approvalRequired: z.boolean().default(false),
  recipients: z.array(z.object({
    name: z.string().min(1),
    email: z.string().email(),
    accountNumber: z.string().min(1),
    bankCode: z.string().optional(),
    routingNumber: z.string().optional(),
    amount: z.number().positive(),
    reference: z.string().optional(),
    metadata: z.record(z.any()).optional()
  })).min(1, "At least one recipient is required"),
  notificationSettings: z.object({
    sendToRecipients: z.boolean().default(true),
    sendToCreator: z.boolean().default(true),
    emailTemplate: z.string().optional()
  }).optional(),
  metadata: z.record(z.any()).optional()
});

const updateBulkPaymentSchema = z.object({
  status: z.enum(['scheduled', 'processing', 'paused', 'cancelled']).optional(),
  scheduledDate: z.string().datetime().optional(),
  description: z.string().optional(),
  notificationSettings: z.object({
    sendToRecipients: z.boolean(),
    sendToCreator: z.boolean(),
    emailTemplate: z.string().optional()
  }).optional()
});

const bulkPaymentProcessSchema = z.object({
  batchId: z.string().uuid(),
  action: z.enum(['start', 'pause', 'resume', 'cancel']),
  reason: z.string().optional(),
  forceOverride: z.boolean().optional()
});

// Mock bulk payment data generator
const generateMockBulkPayment = (id: string, overrides: Partial<any> = {}) => ({
  id,
  name: `Payment Batch ${id.slice(-4)}`,
  description: `Bulk payment processing for ${overrides.name || 'various recipients'}`,
  status: overrides.status || ['draft', 'scheduled', 'processing', 'completed', 'failed'][Math.floor(Math.random() * 5)],
  currency: 'USD',
  totalRecipients: Math.floor(Math.random() * 100) + 10,
  totalAmount: Math.floor(Math.random() * 500000) + 50000,
  processedCount: 0,
  successCount: 0,
  failedCount: 0,
  pendingCount: 0,
  
  timing: {
    createdAt: new Date(Date.now() - Math.random() * 30 * 24 * 60 * 60 * 1000).toISOString(),
    scheduledAt: overrides.scheduledDate || new Date(Date.now() + Math.random() * 7 * 24 * 60 * 60 * 1000).toISOString(),
    startedAt: null,
    completedAt: null,
    estimatedDuration: Math.floor(Math.random() * 3600) + 300 // 5 minutes to 1 hour
  },

  creator: {
    id: 'user_' + Math.random().toString(36).substring(2, 15),
    name: 'John Admin',
    email: 'john.admin@company.com'
  },

  approvals: {
    required: overrides.approvalRequired || Math.random() > 0.7,
    status: 'pending',
    approvers: [
      {
        id: 'approver_1',
        name: 'Jane Manager',
        email: 'jane.manager@company.com',
        role: 'finance_manager',
        status: 'pending'
      }
    ]
  },

  processing: {
    batchSize: 50, // Process in batches of 50
    retryAttempts: 3,
    retryDelay: 30, // seconds
    parallelProcessing: true,
    maxConcurrency: 10
  },

  notifications: {
    sendToRecipients: true,
    sendToCreator: true,
    emailTemplate: 'default',
    webhookUrl: 'https://company.com/webhooks/bulk-payments'
  },

  summary: {
    totalFees: Math.floor(Math.random() * 5000) + 500,
    exchangeRate: overrides.currency !== 'USD' ? 1.1 : 1.0,
    processingTime: null,
    errorRate: 0,
    averageAmount: 0
  },

  compliance: {
    amlChecked: true,
    sanctionsChecked: true,
    fraudChecked: true,
    complianceScore: 95 + Math.random() * 5
  },

  files: [
    {
      id: 'file_1',
      name: 'recipients_list.csv',
      size: 25600,
      uploadedAt: new Date(Date.now() - 2 * 60 * 60 * 1000).toISOString(),
      status: 'processed'
    }
  ],

  metadata: overrides.metadata || {},
  ...overrides
});

const generateMockRecipients = (batchId: string, count: number = 10) => {
  return Array.from({ length: count }, (_, i) => ({
    id: `recipient_${batchId}_${i + 1}`,
    batchId,
    sequenceNumber: i + 1,
    name: `Recipient ${i + 1}`,
    email: `recipient${i + 1}@example.com`,
    accountNumber: `****${Math.floor(Math.random() * 9000) + 1000}`,
    bankCode: ['CHASE', 'BOA', 'WELLS', 'CITI'][Math.floor(Math.random() * 4)],
    routingNumber: '021000021',
    amount: Math.floor(Math.random() * 5000) + 100,
    currency: 'USD',
    reference: `REF_${batchId.slice(-4)}_${String(i + 1).padStart(3, '0')}`,
    status: ['pending', 'processing', 'completed', 'failed'][Math.floor(Math.random() * 4)],
    
    processing: {
      attempts: Math.floor(Math.random() * 3),
      lastAttempt: new Date(Date.now() - Math.random() * 24 * 60 * 60 * 1000).toISOString(),
      estimatedCompletion: new Date(Date.now() + Math.random() * 24 * 60 * 60 * 1000).toISOString(),
      processingTime: Math.floor(Math.random() * 300) + 30 // 30 seconds to 5 minutes
    },

    transaction: {
      id: `txn_${Math.random().toString(36).substring(2, 15)}`,
      status: 'pending',
      fees: Math.floor(Math.random() * 10) + 2,
      exchangeRate: 1.0,
      completedAt: null
    },

    verification: {
      accountVerified: Math.random() > 0.1,
      nameMatch: Math.random() > 0.05,
      riskScore: Math.random() * 0.3, // Low risk
      complianceStatus: 'approved'
    },

    notifications: {
      emailSent: Math.random() > 0.2,
      smsSent: false,
      lastNotification: new Date(Date.now() - Math.random() * 12 * 60 * 60 * 1000).toISOString()
    },

    errors: Math.random() > 0.9 ? [
      {
        code: 'INVALID_ACCOUNT',
        message: 'Bank account number is invalid',
        timestamp: new Date().toISOString(),
        retryable: false
      }
    ] : [],

    metadata: {
      department: ['HR', 'Sales', 'Engineering', 'Marketing'][Math.floor(Math.random() * 4)],
      employeeId: `EMP_${Math.floor(Math.random() * 9000) + 1000}`,
      payrollPeriod: '2024-01'
    }
  }));
};

// GET /api/enterprise/bulk-payments - List bulk payments
export async function GET(request: NextRequest) {
  try {
    const { searchParams } = new URL(request.url);
    const params = Object.fromEntries(searchParams);
    
    const validation = bulkPaymentQuerySchema.safeParse(params);
    if (!validation.success) {
      return NextResponse.json(
        {
          success: false,
          error: 'Invalid query parameters',
          details: validation.error.errors
        },
        { status: 400 }
      );
    }

    const { 
      status, 
      page = '1', 
      limit = '10', 
      sortBy = 'created', 
      sortOrder = 'desc',
      search 
    } = validation.data;

    const pageNum = parseInt(page);
    const limitNum = parseInt(limit);

    // Generate mock bulk payments
    const totalPayments = 47;
    const bulkPayments = Array.from({ length: Math.min(limitNum, 10) }, (_, i) => 
      generateMockBulkPayment(`bulk_${pageNum}_${i + 1}`, { 
        status: status || undefined 
      })
    );

    // Apply search filter
    let filteredPayments = bulkPayments;
    if (search) {
      filteredPayments = bulkPayments.filter(payment => 
        payment.name.toLowerCase().includes(search.toLowerCase()) ||
        payment.description.toLowerCase().includes(search.toLowerCase())
      );
    }

    // Calculate statistics
    const statistics = {
      total: totalPayments,
      byStatus: {
        draft: 5,
        scheduled: 12,
        processing: 3,
        completed: 25,
        failed: 1,
        paused: 1
      },
      totalAmount: 2456789.50,
      totalRecipients: 1247,
      averageAmount: 1970.45,
      successRate: 97.2
    };

    const totalPages = Math.ceil(totalPayments / limitNum);

    return NextResponse.json({
      success: true,
      data: {
        bulkPayments: filteredPayments,
        pagination: {
          page: pageNum,
          limit: limitNum,
          total: totalPayments,
          totalPages,
          hasNext: pageNum < totalPages,
          hasPrev: pageNum > 1
        },
        statistics,
        filters: {
          status,
          search,
          sortBy,
          sortOrder
        }
      }
    });

  } catch (error) {
    console.error('Bulk payments list error:', error);
    return NextResponse.json(
      {
        success: false,
        error: 'Failed to fetch bulk payments'
      },
      { status: 500 }
    );
  }
}

// POST /api/enterprise/bulk-payments - Create bulk payment
export async function POST(request: NextRequest) {
  try {
    const body = await request.json();
    
    const validation = createBulkPaymentSchema.safeParse(body);
    if (!validation.success) {
      return NextResponse.json(
        {
          success: false,
          error: 'Invalid bulk payment data',
          details: validation.error.errors
        },
        { status: 400 }
      );
    }

    const paymentData = validation.data;
    
    // Calculate totals
    const totalAmount = paymentData.recipients.reduce((sum, recipient) => sum + recipient.amount, 0);
    const totalFees = paymentData.recipients.length * 2.5; // $2.50 per recipient
    const estimatedDuration = Math.ceil(paymentData.recipients.length / 50) * 300; // 5 minutes per batch of 50

    const newBulkPayment = {
      id: `bulk_${Date.now()}`,
      ...paymentData,
      status: paymentData.approvalRequired ? 'draft' : 'scheduled',
      totalRecipients: paymentData.recipients.length,
      totalAmount,
      processedCount: 0,
      successCount: 0,
      failedCount: 0,
      pendingCount: paymentData.recipients.length,
      
      timing: {
        createdAt: new Date().toISOString(),
        scheduledAt: paymentData.scheduledDate || new Date(Date.now() + 24 * 60 * 60 * 1000).toISOString(),
        startedAt: null,
        completedAt: null,
        estimatedDuration
      },

      creator: {
        id: 'current_user',
        name: 'Current User',
        email: 'user@company.com'
      },

      approvals: {
        required: paymentData.approvalRequired,
        status: paymentData.approvalRequired ? 'pending' : 'auto_approved',
        approvers: paymentData.approvalRequired ? [
          {
            id: 'approver_1',
            name: 'Finance Manager',
            email: 'finance@company.com',
            role: 'finance_manager',
            status: 'pending'
          }
        ] : []
      },

      processing: {
        batchSize: 50,
        retryAttempts: 3,
        retryDelay: 30,
        parallelProcessing: true,
        maxConcurrency: 10
      },

      notifications: paymentData.notificationSettings || {
        sendToRecipients: true,
        sendToCreator: true,
        emailTemplate: 'default'
      },

      summary: {
        totalFees,
        exchangeRate: 1.0,
        processingTime: null,
        errorRate: 0,
        averageAmount: totalAmount / paymentData.recipients.length
      },

      compliance: {
        amlChecked: false,
        sanctionsChecked: false,
        fraudChecked: false,
        complianceScore: null
      },

      files: [],
      metadata: paymentData.metadata || {}
    };

    // In a real implementation, save to database and trigger compliance checks
    // await saveBulkPayment(newBulkPayment);
    // await triggerComplianceChecks(newBulkPayment.id);

    return NextResponse.json({
      success: true,
      message: 'Bulk payment batch created successfully',
      data: {
        bulkPayment: newBulkPayment,
        nextSteps: paymentData.approvalRequired ? [
          'Awaiting approval from finance manager',
          'Compliance checks will be performed after approval',
          'Payment will be processed on scheduled date'
        ] : [
          'Compliance checks initiated',
          'Payment scheduled for processing',
          'Recipients will be notified before processing'
        ]
      }
    }, { status: 201 });

  } catch (error) {
    console.error('Bulk payment creation error:', error);
    return NextResponse.json(
      {
        success: false,
        error: 'Failed to create bulk payment batch'
      },
      { status: 500 }
    );
  }
}

// PUT /api/enterprise/bulk-payments/[id] - Update bulk payment
export async function PUT(request: NextRequest) {
  try {
    const url = new URL(request.url);
    const batchId = url.pathname.split('/').pop();
    
    if (!batchId) {
      return NextResponse.json(
        {
          success: false,
          error: 'Batch ID is required'
        },
        { status: 400 }
      );
    }

    const body = await request.json();
    
    const validation = updateBulkPaymentSchema.safeParse(body);
    if (!validation.success) {
      return NextResponse.json(
        {
          success: false,
          error: 'Invalid update data',
          details: validation.error.errors
        },
        { status: 400 }
      );
    }

    // In a real implementation, fetch and update the bulk payment
    const existingPayment = generateMockBulkPayment(batchId);
    
    const updatedPayment = {
      ...existingPayment,
      ...validation.data,
      updatedAt: new Date().toISOString()
    };

    return NextResponse.json({
      success: true,
      message: 'Bulk payment batch updated successfully',
      data: {
        bulkPayment: updatedPayment
      }
    });

  } catch (error) {
    console.error('Bulk payment update error:', error);
    return NextResponse.json(
      {
        success: false,
        error: 'Failed to update bulk payment batch'
      },
      { status: 500 }
    );
  }
}

// POST /api/enterprise/bulk-payments/process - Process bulk payment
export async function PATCH(request: NextRequest) {
  try {
    const body = await request.json();
    
    const validation = bulkPaymentProcessSchema.safeParse(body);
    if (!validation.success) {
      return NextResponse.json(
        {
          success: false,
          error: 'Invalid process request',
          details: validation.error.errors
        },
        { status: 400 }
      );
    }

    const { batchId, action, reason, forceOverride } = validation.data;

    // In a real implementation, process the batch action
    const result = {
      batchId,
      action,
      status: 'success',
      message: `Bulk payment batch ${action} initiated successfully`,
      timestamp: new Date().toISOString(),
      estimatedCompletion: action === 'start' ? 
        new Date(Date.now() + 30 * 60 * 1000).toISOString() : null
    };

    if (reason) {
      result.message += ` (Reason: ${reason})`;
    }

    return NextResponse.json({
      success: true,
      data: result
    });

  } catch (error) {
    console.error('Bulk payment process error:', error);
    return NextResponse.json(
      {
        success: false,
        error: 'Failed to process bulk payment batch'
      },
      { status: 500 }
    );
  }
}

// GET /api/enterprise/bulk-payments/[id]/recipients - Get batch recipients
export async function GET_RECIPIENTS(request: NextRequest) {
  try {
    const url = new URL(request.url);
    const batchId = url.pathname.split('/')[4]; // Extract batch ID from path
    
    if (!batchId) {
      return NextResponse.json(
        {
          success: false,
          error: 'Batch ID is required'
        },
        { status: 400 }
      );
    }

    const recipients = generateMockRecipients(batchId, 25);

    return NextResponse.json({
      success: true,
      data: {
        batchId,
        recipients,
        summary: {
          total: recipients.length,
          pending: recipients.filter(r => r.status === 'pending').length,
          processing: recipients.filter(r => r.status === 'processing').length,
          completed: recipients.filter(r => r.status === 'completed').length,
          failed: recipients.filter(r => r.status === 'failed').length
        }
      }
    });

  } catch (error) {
    console.error('Bulk payment recipients error:', error);
    return NextResponse.json(
      {
        success: false,
        error: 'Failed to fetch batch recipients'
      },
      { status: 500 }
    );
  }
}