# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PayPass API
  description: |
    PayPass is a comprehensive payment platform providing secure, scalable payment processing solutions.
    
    ## Features
    - Payment Processing
    - Wallet Management  
    - User Authentication
    - Enterprise Billing
    - Fraud Detection
    - Compliance Management
    
    ## Authentication
    PayPass API uses API key authentication. Include your API key in the Authorization header:
    ```
    Authorization: Bearer YOUR_API_KEY
    ```
    
    ## Rate Limiting
    API requests are rate limited based on your subscription plan:
    - Basic: 100 requests/hour
    - Premium: 1000 requests/hour  
    - Enterprise: 10000 requests/hour
    
    ## Webhooks
    PayPass can send webhook notifications for various events. Configure webhooks in your dashboard.
    
  version: 2.1.0
  contact:
    name: PayPass API Support
    email: api-support@paypass.com
    url: https://docs.paypass.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://paypass.com/terms
servers:
  - url: https://api.paypass.com/v1
    description: Production server
  - url: https://api-staging.paypass.com/v1
    description: Staging server
  - url: https://api-dev.paypass.com/v1
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /payments:
    post:
      tags:
        - Payments
      summary: Create a payment
      description: Process a new payment transaction
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

    get:
      tags:
        - Payments
      summary: List payments
      description: Retrieve a list of payments
      operationId: listPayments
      parameters:
        - name: limit
          in: query
          description: Number of payments to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of payments to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: status
          in: query
          description: Filter by payment status
          schema:
            type: string
            enum: [pending, completed, failed, cancelled]
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /payments/{paymentId}:
    get:
      tags:
        - Payments
      summary: Get payment by ID
      description: Retrieve a specific payment by its ID
      operationId: getPayment
      parameters:
        - name: paymentId
          in: path
          required: true
          description: Payment ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          $ref: '#/components/responses/NotFound'

  # User Endpoints
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreatePaymentRequest:
      type: object
      required:
        - amount
        - currency
        - source
        - destination
      properties:
        amount:
          type: number
          format: decimal
          minimum: 0.01
          description: Payment amount
          example: 100.50
        currency:
          type: string
          length: 3
          description: ISO 4217 currency code
          example: USD
        source:
          $ref: '#/components/schemas/PaymentSource'
        destination:
          $ref: '#/components/schemas/PaymentDestination'
        description:
          type: string
          maxLength: 255
          description: Payment description
          example: Payment for services
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Payment ID
        amount:
          type: number
          format: decimal
          description: Payment amount
        currency:
          type: string
          description: Payment currency
        status:
          type: string
          enum: [pending, completed, failed, cancelled]
          description: Payment status
        source:
          $ref: '#/components/schemas/PaymentSource'
        destination:
          $ref: '#/components/schemas/PaymentDestination'
        description:
          type: string
          description: Payment description
        fees:
          type: number
          format: decimal
          description: Processing fees
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    PaymentSource:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [wallet, bank_account, card]
          description: Source type
        wallet_id:
          type: string
          format: uuid
          description: Wallet ID (if type is wallet)
        account_number:
          type: string
          description: Account number (if type is bank_account)
        card_token:
          type: string
          description: Card token (if type is card)
    PaymentDestination:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [wallet, bank_account, card]
          description: Destination type
        wallet_id:
          type: string
          format: uuid
          description: Wallet ID (if type is wallet)
        account_number:
          type: string
          description: Account number (if type is bank_account)

    # User Schemas
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Items skipped
        total:
          type: integer
          description: Total items
        has_more:
          type: boolean
          description: More items available
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        error_code:
          type: string
          description: Error code
        details:
          type: string
          description: Detailed error information
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid request data"
            error_code: "INVALID_REQUEST"
            details: "The amount field is required"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authentication required"
            error_code: "UNAUTHORIZED"
            details: "Valid API key required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            error_code: "NOT_FOUND"
            details: "The requested resource does not exist"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate limit exceeded"
            error_code: "RATE_LIMITED"
            details: "You have exceeded your API rate limit"
...